requirements.txt
pyTelegramBotAPI
google-generativeai git add requirements.txt
